+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - (Feature) (Scheduler) Merge Strategy
 - (Feature) (Networking) Endpoints Destination
 - (Improvement) Improve Metrics Handling
+- (Feature) (Scheduler) Create Integration Profile
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries